CRUD for Shipments Completed
Created:
- shipmentController.js: contains the API calls for CRUD shipments
- shipmentModel.js: defines the fields  for the shipment. Worth noting it has a field with an array of objects.
- shipmentRoutes.js:: defines the routes used for the API calls

Updated:
- server.js: now uses cron (npm i node-cron) to schedule the job that automatically deletes shipments whose date has passed.

Note, the shipments are created with this kind of API call to http://localhost:5001/api/shipments/create
{
  "carrier": "TOBEDELETED",
  "expectedArrival": "2023-07-01T12:00:00Z",
  "trackingNumber": "1Z123445T605271688",
  "itemsContained": [
            "6447788d3eefafb9cf0935e7",
            "64477156d154eba55ba7b4e5"
      ],

}

Take note that the itemsContained actually contains an array of Mongo ObjectIds. Important for when creating them from the front end, we will likely have to pull from the list of items that the user will want to add into the shipment to get their IDs properly.