Deleting an Item Functionality
Pressing the delete button brings up a prompt making sure the user wants to delete the items.
Properly updates the items state, and also properly deletes in the backend the item.

Changes persist after refresh and right after confirmation.
Only thing left for items is edit, but that is less pending.